Plus de tests

+++ src/main/java/org/example/Calculator.java
@@ -13,7 +13,7 @@ public class Calculator {
         return a * b;
     }
 
-    int div(int a, int b) {
+    double div(int a, int b) {
         return a / b;
     }
 }
